[nrf noup] soc: nordic: Support TF-M for poweroff
Enable going to system off when the TF-M API is available. This
calls the relevant TF-M platform API to enter system off mode using
TF-M. This is a noup because the TF-M service for system off is
currently located in sdk-nrf only.

This is meant to be a short lived commit, work is already ongoing
to upstream an official TF-M system off solution that can be be
added to upstream Zephyr.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
(cherry picked from commit 1ccd687ded0cc174e6329fbaf5dda7af5f489a2c)